refactor: Move Tool subclasses into definition.py files
- Move BashTool class from tool.py to execute_bash/definition.py
- Move FileEditorTool class from tool.py to str_replace_editor/definition.py
- Update __init__.py files to import from definition.py instead of tool.py
- Remove separate tool.py files to consolidate code organization
- Add local imports to avoid circular import issues
- All tests still pass and functionality remains unchanged

This follows the cleaner pattern of keeping Tool definitions alongside
their schemas and base tool definitions in a single file.
