+++ openhands/tools/execute_bash/__init__.py
@@ -1,10 +1,10 @@
 from openhands.tools.execute_bash.definition import (
+    BashTool,
     ExecuteBashAction,
     ExecuteBashObservation,
     execute_bash_tool,
 )
 from openhands.tools.execute_bash.impl import BashExecutor
-from openhands.tools.execute_bash.tool import BashTool
 
 
 __all__ = [
