+++ openhands/tools/str_replace_editor/__init__.py
@@ -1,10 +1,10 @@
 from openhands.tools.str_replace_editor.definition import (
+    FileEditorTool,
     StrReplaceEditorAction,
     StrReplaceEditorObservation,
     str_replace_editor_tool,
 )
 from openhands.tools.str_replace_editor.impl import FileEditorExecutor, file_editor
-from openhands.tools.str_replace_editor.tool import FileEditorTool
 
 
 __all__ = [
